requirements.txt
asgiref==3.7.2
Babel==2.14.0
Django==5.0.3
django-environ==0.11.2
django-money==3.4.1
django-phonenumber-field==7.3.0
django-soft-delete==1.0.12
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==24.9.0
phonenumberslite==8.13.32
pillow==10.2.0
psycopg2-binary==2.9.9
py-moneyed==3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
